Save and restore upgrade graph state to a secret
In order to keep state across multiple restarts, periodically
snapshot the upgrade graph state to a Kube secret and restore from
it on startup. The graph is gzipped (which is why we use a secret).
A future commit should prune / sort the graph to make it easier to
restore.

Add `/archive/graph` to let a client dump the graph.