Verify and sign releases as a standalone action
A second release controller can audit and verify the actions of the
first, loading release images, performing verificaton, and then
signing and publishing the contents to a GCS bucket. This gives an
end to end flow for ensuring content is trusted.

New flags:

  --audit URL (file:// or gs://) to target destination to write to
  --audit-gcs-service-account PATH service account file
  --audit-cli-image IMAGE a CLI image that has been trusted to verify
  --sign GPG_KEYRING to sign releases that are audited

Passing --audit changes the mode for the controller and normal release
maintenance won't be performed. Instead, the controller will audit and
report results.