Correctly include other releases as compare targets
When we picked up a newer semver, https://github.com/blang/semver/issues/55
caused other releases to be ignored from the comparison drop down.
For now work around that by double parsing (tolerant and normal)
the release name to see if it's stable. Also include the stable tags
which are also not semver valid.