Adding an "Untested upgrades" section to the release detail page
We received a request to provide a visual indication of upgrades, that
have not been tested, for the release admins to potentually take action
upon (https://issues.redhat.com/browse/DPCR-25).  These changes gather
the list of baked-in upgrades and compares them against the results of the
know upgrade tests.  If any upgrades are not accounted for, they are
listed in a warning box on the release detail page.