bugzilla: workaround for permissions issue with certain bugs
There are some bugs that have an unusual permissions issue that results
in the bugzilla client failing to get external bugs. This commit is a
workaround for that issue that just logs the error instead of returning
it, allowing the release to move into the `bugs-verified` state instead
of looping indefinitely.