Correctly support multiple release namespaces
The controller historically accepted multiple streams but code
was tied to the first version. Replace all usage of the single
stream with the more correct code. For controllers accepting a
single --release-namespace this change is a no-op. The first
consumer will be the audit signer which can easily verify this
code path.