Support --publish-namespace for target namespaces
The current --release-namespace flag treats the first entry specially.
Now that we correctly support multiple release streams, we need to add
a namespace that is watched but not processed for image stream changes
to handle the case where a publish job wants to keep another image stream
in sync (and uses the lister to do so). --publish-namespace will register
a lister on the namespace but not look at it for release changes.