Fixing support for RHCOS linking to architecture specific releases and differences
It was pointed out that the architecture specific links to RHCOS content were
all pointing to the x86_64 streams and therefore presenting incorrect information
to the users.  These changes interrogate the architecture out of the pullSpec
that is passed into the changelog function. Once the archtecture is known, it is
used to create the correct URL to the architecture specific content and RHCOS
difference tool.