Updating the deferred recovery logic
I observed that in most cases there was a "begin" message, but never a corresponding
"end" message surrounding the various sync loops that occur.  It appears that this was
caused by the deferal of the recover() method followed, immediately, by a panic(nil).
I have updated the offending logic to only issue the panic() call when an error is
returned from the recover() logic.