cmd: add support for running periodics in release-controller
This PR adds support for the release-controller to run periodics defined
in release configs, similar to verify jobs. It uses the test-infra cron
pkg and has a similar structure to prow's horologium sync function, but
modified to use the data structures, interfaces, and configuration that
the release-controller has.

To add a periodic job, a user would add the job to the `periodic`
section of the release config with the name of the job, the interval or
cron timing the user wants, and an `upgrade` bool and `upgradeFrom`
string if it is an upgrade job.

Previously, release informers (specifically upgrades) were run by
horologium and used curl to get release tags and annotated themselves to
be picked up by release-controller. This PR allows us to fix that
problem by creating the prowjobs directly from release-controller with
the proper annotations and env vars from the start.