cmd: add config validation mode
This PR adds a config validation mode to the release-controller that can
be run with the `--validate-configs` flag, which takes the directory
containing the configs as its value. It is intended to be run as a
presubmit in `openshift/release`. At the moment, it verifies that all
the release configs can be unmarshalled and does verification on the
periodic job configs. More verification functions can be added to it in
the future to validate that other parts of the release configs are valid
as well.