Files-cache messaging format fix
Someone reached out in slack stating that they were unable to download the
installer for a specific release.  Upon further investigation, it turns out
that there was a bug in the FileServer logic when handling the exceptions that
allow for the retrying of the downloads.  The specific error was:

File "/tmp/serve.py", line 67, in do_GET
  print("Unable to get release tools for %s: %s" % name, e.output)
TypeError: not enough arguments for format string

The fix is to enclose the values in parenthesis.