Files-cache exception handling fix
After fix #228 was deployed, the files-cache was able to make more progress
downloading installers.  Unfortunately, the logic that interrogates the errors
from oc, is using the old (python2) style strings and the errors that are
returned are binary strings.  This PR updates the logic to compare the strings
as binary and also adds another check for an updated message format.