validate: add validation for defined release-controller jobs
A new field has been added to ci-operator that adds a label to jobs to
define them as release-controller jobs. This reduces confusion when
defining jobs in ci-operator configs that are supposed to be executed
periodically or as verification by release-controller.

This PR adds validation to make sure that all jobs defined as periodics
or verification in release-controller have the correct label on the job
config, and all job configs with the label are exist in the
release-controller configs.