sync_verify_prow: always set an initial release
This PR fixes an issue where ci-operator would fail to start a job as it
would try to pull a non-existant release based on the `releases`
configured in the ci-operator config. This situation could occur on
creation of a new release or when many tags fail to be accepted and only
1 accepted tag exists for a release. To fix this, we configure
  release-controller to always set the initial release image to override
what is defined in the ci-operator config.