main: do not reload on in cluster token change
The client used by the release-controller for in-cluster configs is set
up to automatically reload in-cluster configs periodically without
needing to change how it is currently configured in the
release-controller. This means that the watch on the token file's path
that triggered periodic restarts when the bound token was updated can be
removed.