bugzilla: use full email for bz comment checks
When this code was first made, I checked the `creator` value by curling
a public bug that the bugzilla bot had made a public comment on. The
creator was set as `openshift-bugzilla-robot`. However, a couple of bugs
with multiple PRs have had multiple identical comments posted on them
recently. Apparently, when the REST call is performed with
authentication, the creator is `openshift-bugzilla-robot@redhat.com`.