bugzilla: improvements to auto-verification for QE
QE has recently asked to simplify the requirements for the bugzilla
autoverification back to just the LGTM by QA contact requirement.
Another request was to clarify that, after this message is sent, the QA
contact must manually move the bug to `VERIFIED` themselves.

This resolves OCPCRT-110.

I also added the release with the bugfix to the comment, as I believe
that could be useful and it is a simple addition.