release-controller-api: clone http into separate binary
This PR creates a new command called `release-controller-api` that is a
stripped down version of the `release-controller` that contains only the
user interface and API components of the controller. The http code will
be removed from the main `release-controller` binary in a future PR
after the new `release-controller-api` service is up and running. Any
code that was shared between the http and non-http code has been moved
into files in `pkg/release-controller`.