Include `phase` field in the release info API
Currently ART is using the `/api/v1/releasestream/{release}/latest` API to determine whether the release being promoted is accepted by the release controller.
This method has some timing issues when ART attempts to rerun some automation for an old release.
It would be very convenient to query the `Phase` field for a specific release with the `/api/v1/releasestream/{release}/release/{tag}` API.
With this power, ART's automation will be smarter and idempotent.