Remove unhelpful warnings from UI The two warnings being removed are essentially permanently displayed on the release controller today.
'Some tags are significantly older ...' - This is expected as upstream
changes slow. CI may rebuild images for various reasons, but ART
will not. This is permanent in the case where a release goes EOL.

'No downstream builds have been pushed ...' - ART's imagestream
is carefully curated and pruned with automation. It is limited
to release payload components. The CI imagestream is not pruned
and has collected images unrelated to the release payload.