Temporary fix to allow bugzilla verification to proceed
When we introduced the Jira integration logic there were issues with the
integration with `oc` that are preventing us from being able to process
Bugs.  This fix reverts us back to the "old" logic and should allow for
the verification logic to resume.

This PR can be reverted if/when an appropriate fix is made in `oc`.