main: add secret agent to fix bugzilla authentication
This PR makes the main function add bugzilla and github secrets to the
secret agent. While some flagutils add the secret to the agent before
creating the client, bugzilla doesn't, which causes auth to not work at
all. This PR fixes that by adding all tokens that are not added to the
secretAgent automatically.