pkg/jira: ignore issues not ON_QA
This PR changes the behavior of the jira verifier to ignore all issues
that are not marked as `ON_QA` in jira. This previous behavior was
legacy functionality that was done due to an issue in ART's automation
that would mark bugs in Bugzilla as `ON_QA` very late, causing the issue
to be ignored by the `release-controller`. The release-controller would
make a comment so it was clear that the problem was not from the
`release-controller` side. A side-effect is that this can cause the
`release-controller` to make incorrect comments on bugs where not all
PRs have merged. This shouldn't be an issue anymore, so the
`release-controller` will now assume that all bugs are marked as `ON_QA`
promptly.