http: add collapse support for all imagestreams (#595)
* http: add collapse support for all imagestreams

This PR adds support for collapsable rows for all imagestreams on the
homepage of the `release-controller-api` and adds new pages for each
individual imagestream at `/releasestream/releaseConfigName` for users
who wish to see the upgrade graph. Due to the massive size of the
`4-stable` upgrade graph, the upgrade graph is omitted for
`/releasestream/4-stable`.

* http: fix bugs found in review