release-controller-api: print team acceptance status
This adds a new column to the release streams called "Team Approvals"
which list all teams that have accepted or rejected a release payload
via a label with a key in the format of "...teamName_state" which can
have a value of "Accepted" or "Rejected". This also prints the list of
the tag's page above the test results.