rhcos: handle new RHEL-based version strings
As part of https://github.com/openshift/enhancements/pull/1637, the
version string for RHCOS will change from being OCP+RHEL-based (e.g.
419.96...) to being purely RHEL-based (e.g. 9.6...).

Adapt the logic for this new scheme so that it links to the right
stream.

We should eventually clean this up though so that the stream name is
available more directly to the release controller so that it doesn't
need to do any guessing.