release-controller-api: render RPM diff
This basically fixes the gap that the infobox added in #648 warns about.
Instead of relying on the CoreOS release browser differ, we directly
query for the diff using `oc adm release info`, which recently learned
to do this, and render it at the bottom of the page. The `diff` link now
simply links to that section of the page.

This removes the VPN requirement on being able to see the RPM diff.

Only do this for the new layered node images, as well as for 4.19.
We could in theory do this all the way back to e.g. 4.12 when RHCOS
switched to "OSTree native containers", but let's be conservative. This
provide a natural rollout of the feature in case it needs ironing out.

Implementation-wise, we don't want to block the main changelog render on
getting the RPM diff. So we structure this the same way as the changelog
render itself; if the `oc` command takes too long, we display a nice
message to check again later. Once the command finishes and the output
is cached, we include it in the following render requests.