release-controller-api: render RPM diff
This basically fixes the gap that the infobox added in #648 warns about.
Instead of relying on the CoreOS release browser differ, we directly
query for the diff using `oc adm release info`, which recently learned
to do this, and render it at the bottom of the page.

This removes the VPN requirement on being able to see the RPM diff.

In the case of the new layered node images, replace the diff link to
instead link to this new rendered section. We could in theory always
link to that, but let's be conservative. This provide a natural rollout
of the feature in case it needs ironing out.

Implementation-wise, we don't want to block the main changelog render on
getting the RPM diff. So we structure this the same way as the changelog
render itself; if the `oc` command takes too long, we display a nice
message to check again later. Once the command finishes and the output
is cached, we include it in the following render requests.