release-mirror-cleanup-controller: fix logic
Make the release-mirror-cleanup-controller work correctly. In the initial
version, the code assumed that the imagestreams with the release config
specifying the alternate mirror repo also contained the release tags itself.
This is wrong, and release tags are kept in imagestreams whose release
definition contains `"as":"Stable"`.

The new logic now generates 2 sets when it loops through the streams: the names
of the tags in releases whose definition contains `"as":"Stable"` and the names
of the alternate repos found in other imagestreams with a release definition
annotation. We then loop through all alternate repos, and any tag in the
alternate repos that isn't part of the `validTags` set is removed.