http_changelog: don't add `target` attribute to anchor links
The CoreOS diff is a link to a separate section on the same page. It's
weird to have that open a new page instead of just scrolling the current
page.

Change it.

There are no other in-page anchor links that I can see, but if there
were/would be in the future, I think we'd probably want this same
behaviour, so we don't need to special-case anything here.