release_info: add helper to run `oc` commands
It's silly that we have to lookup the `oc` path each time. Let's just do
it once and cache that.

Also, there's a lot of boilerplate around running commands and capturing
output. Factor that out into a function.

Move over two functions to use this new helper to get the ball rolling.
Leave markers on the other callsites where it could be used for future
refactor.

The `ocCmdExt()` that takes a directory is not currently needed but will
be in a future patch.