rhcos: remove base image diff link, add infobox
There's a lot of confusion with the new layered model and where to look
for package diffs and lists.

First, stop adding a diff link on the CoreOS components line, even
if it's to go to the node image section lower down. It's confusing and
not correct because the components info is always about the base image,
not the node image itself.

Keep linkifying to CoreOS version strings that go to the CoreOS release
browser, but also add an infobox that makes it clear that the info there
is about the _base_ image and what 95% of viewers want instead is the
node image info lower down, which we then link to.

Include this infobox in both the CoreOS upgrade and no-upgrade case.

Even if the base layer didn't change, the node image may still have
changed and contain a diff of OpenShift RPMs. So emitting the box in
that case too also implicitly sets the `#node-image-info` marker that
enables RPM diff querying.

Finally, in the "Node Image Info" section, put the base version
info (which links to the CoreOS release browser) at the bottom to
de-emphasize it.