rhcos: also display extensions list
This is the final big piece that was useful from the previous CoreOS
release browser. Users want to be able to know what extensions landed
in a release.

Querying this is not driven through to `oc` like the node image rpmdb
is, but the metadata for it can also be cheaply queried using `oc`:

https://github.com/openshift/os/commit/31816acb1ae377c9c48f1e4bc70fbf63cf4adc2d

Query it and render it.