rhcos: various fixes and enhancements (#691)
* rhcos: fix regexes to match both 419.x and 9.x

The current regexes don't match the new 9.x versioning scheme used by
the CoreOS base images used to build the node image. There's no need to
isolate the digits here, so just simplify it so it matches both.

* rhcos: add newline between diff sections in Markdown

Otherwise, we would have e.g.

    ### Changed

    * foo x -> y
    ### Added

Which changes the meaning in Markdown and that would then make the diff
not render properly.

* http_changelog: don't add `target` attribute to anchor links

The CoreOS diff is a link to a separate section on the same page. It's
weird to have that open a new page instead of just scrolling the current
page.

Change it.

There are no other in-page anchor links that I can see, but if there
were/would be in the future, I think we'd probably want this same
behaviour, so we don't need to special-case anything here.

* release_info: add helper to run `oc` commands

It's silly that we have to lookup the `oc` path each time. Let's just do
it once and cache that.

Also, there's a lot of boilerplate around running commands and capturing
output. Factor that out into a function.

Move over two functions to use this new helper to get the ball rolling.
Leave markers on the other callsites where it could be used for future
refactor.

The `ocCmdExt()` that takes a directory is not currently needed but will
be in a future patch.

* http_changelog: change "CoreOS Package Diff" to "Node Image Info"

This makes the added section more generic. The package diff is still
currently the only thing rendered there, but that soon won't be the case
and so we need better naming and wording.

* rhcos: remove base image diff link, add infobox

There's a lot of confusion with the new layered model and where to look
for package diffs and lists.

First, stop adding a diff link on the CoreOS components line, even
if it's to go to the node image section lower down. It's confusing and
not correct because the components info is always about the base image,
not the node image itself.

Keep linkifying to CoreOS version strings that go to the CoreOS release
browser, but also add an infobox that makes it clear that the info there
is about the _base_ image and what 95% of viewers want instead is the
node image info lower down, which we then link to.

Include this infobox in both the CoreOS upgrade and no-upgrade case.

Even if the base layer didn't change, the node image may still have
changed and contain a diff of OpenShift RPMs. So emitting the box in
that case too also implicitly sets the `#node-image-info` marker that
enables RPM diff querying.

Finally, in the "Node Image Info" section, put the base version
info (which links to the CoreOS release browser) at the bottom to
de-emphasize it.

* rhcos: also display node image package list

Right now, we only display diffs, but it's useful to know the full
package set too for a given node image.

This is already supported by `oc`, and reuses the same cached data as
that used for calculating diffs. We just need to plumb it through.

* rhcos: also display extensions list

This is the final big piece that was useful from the previous CoreOS
release browser. Users want to be able to know what extensions landed
in a release.

Querying this is not driven through to `oc` like the node image rpmdb
is, but the metadata for it can also be cheaply queried using `oc`:

https://github.com/openshift/os/commit/31816acb1ae377c9c48f1e4bc70fbf63cf4adc2d

Query it and render it.