http_changelog: hold render for node image info
Use the same trick for the node image info section that that main
changelog section uses to first display a "loading..." box, flush, then
hide the box once the render is complete.

This change avoids users having to refresh the page for the node image
info section when accessing a release page that isn't yet cached,
without sacrificing how long it takes to have the main changelog section
rendered.

We still put a hard timeout of 15 seconds to render it. Even on a cold
startup, it should normally be much faster than the changelog case which
requires massive git cloning.