http_changelog: simplify node image info render logic
Change the loading alert to an `alert-info` so it renders in blue to
match the changelog one.

Always flush after printing the "Node Image Info" heading so that the
anchor is valid earlier during the render. Also, don't wait 1s to print
the info alert so users can tell earlier that work is happening. If the
result is cached, the box will never be visible anyway since it gets
hidden so quickly.